pyproject.toml
[project]
name = "livephish"
version = "0.1.0"
description = "Interactive LivePhish show browser and downloader"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "InquirerPy>=0.3.4",
    "rich>=13.0",
    "click>=8.1",
    "mutagen>=1.47",
    "keyring>=25.0",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
]

[project.scripts]
livephish = "livephish.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/livephish"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "pyinstaller>=6.0",
]
